Rename `messages` module to `message`
